+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.0.23
+
+- Location selection for http checks (can be enabled via STEADYBIT_EXTENSION_ENABLE_LOCATION_SELECTION env var, requires platform => 2.1.27)
+- Use "error" in the expected HTTP status code field to verify that requests are returning an error
+- Use uid instead of name for user statement in Dockerfile
+
 ## v1.0.22
 
 - Update dependencies (go 1.23)